DDF-1346 must pass BigInteger to BigInteger.equals()

+++ spatial/wfs/2.0.0/spatial-wfs-v2_0_0-source/src/main/java/org/codice/ddf/spatial/ogc/wfs/v2_0_0/catalog/source/WfsSource.java
@@ -640,7 +640,7 @@ public class WfsSource extends MaskableImpl implements FederatedSource, Connecte
 
             if (featureCollection.getNumberReturned() == null) {
                 LOGGER.warn("Number Returned Attribute was not added to the response");
-            } else if (!featureCollection.getNumberReturned().equals(numResults)) {
+            } else if (!featureCollection.getNumberReturned().equals(BigInteger.valueOf(numResults))) {
                 LOGGER.warn(
                         "Number Returned Attribute ({}) did not match actual number returned ({})",
                         featureCollection.getNumberReturned(), numResults);
